Memory mgmt: define per-class memory management strategies, also update testing for __wx_destroyed__

+++ wxruby2/swig/classes/include/wxSizer.h
@@ -94,7 +94,7 @@ public:
 
     virtual bool Detach( wxWindow *window );
     virtual bool Detach( wxSizer *sizer );
-    virtual bool Detach( int index );
+    virtual bool Detach( size_t index );
 
     virtual void Clear( bool delete_windows = false );
     virtual void DeleteWindows();
